PositionComponent set and constructors now use floats for x,y. fix moving non physics based entities in PhysicsSystem

+++ core/src/main/java/com/redsponge/redengine/screen/components/PositionComponent.java
@@ -19,11 +19,11 @@ public class PositionComponent implements Component {
         this(0, 0, 0);
     }
 
-    public PositionComponent(int x, int y) {
+    public PositionComponent(float x, float y) {
         this(x, y, 0);
     }
 
-    public PositionComponent(int x, int y, int z) {
+    public PositionComponent(float x, float y, int z) {
         this.x = x;
         this.y = y;
         this.z = z;
@@ -59,14 +59,14 @@ public class PositionComponent implements Component {
         return this;
     }
 
-    public PositionComponent set(int x, int y) {
+    public PositionComponent set(float x, float y) {
         this.x = x;
         this.y = y;
         this.beenSet = true;
         return this;
     }
 
-    public PositionComponent set(int x, int y, int z) {
+    public PositionComponent set(float x, float y, int z) {
         this.x = x;
         this.y = y;
         this.z = z;
